moving tests from TG to T3K
