+++ tests/ttnn/distributed/test_multidevice_TG.py
@@ -1474,116 +1474,6 @@ def test_device_line_all_gather_8x4_data(mesh_device, cluster_axis: int, dim: in
     # )
 
 
-@pytest.mark.parametrize(
-    "mesh_device",
-    [
-        pytest.param((4, 8), id="4x8_grid"),
-        pytest.param((1, 32), id="1x32_grid"),
-    ],
-    indirect=True,
-)
-def test_visualize_tensor_col_sharded(mesh_device):
-    rows, cols = mesh_device.shape
-    tile_size = 32
-    full_tensor = torch.rand((1, 1, tile_size * rows, tile_size * cols), dtype=torch.bfloat16)
-    mesh_mapper = ttnn.create_mesh_mapper(
-        mesh_device,
-        ttnn.MeshMapperConfig(
-            placements=[
-                ttnn.PlacementReplicate(),
-                ttnn.PlacementShard(3),
-            ],
-        ),
-    )
-    ttnn_tensor = ttnn.from_torch(full_tensor, mesh_mapper=mesh_mapper, layout=ttnn.Layout.ROW_MAJOR)
-    ttnn.visualize_tensor(ttnn_tensor)
-    ttnn_tensor = ttnn_tensor.to(mesh_device)
-    ttnn.visualize_tensor(ttnn_tensor)
-
-
-@pytest.mark.parametrize(
-    "mesh_device",
-    [
-        pytest.param((4, 8), id="4x8_grid"),
-        pytest.param((1, 32), id="1x32_grid"),
-    ],
-    indirect=True,
-)
-def test_visualize_tensor_row_sharded(mesh_device):
-    rows, cols = mesh_device.shape
-    tile_size = 32
-    full_tensor = torch.rand((1, 1, tile_size * rows, tile_size * cols), dtype=torch.bfloat16)
-    mesh_mapper = ttnn.create_mesh_mapper(
-        mesh_device,
-        ttnn.MeshMapperConfig(
-            placements=[
-                ttnn.PlacementShard(3),
-                ttnn.PlacementReplicate(),
-            ],
-        ),
-    )
-    ttnn_tensor = ttnn.from_torch(full_tensor, mesh_mapper=mesh_mapper, layout=ttnn.Layout.ROW_MAJOR)
-    ttnn.visualize_tensor(ttnn_tensor)
-    ttnn_tensor = ttnn_tensor.to(mesh_device)
-    ttnn.visualize_tensor(ttnn_tensor)
-
-
-@pytest.mark.parametrize("mesh_device", [pytest.param((4, 8), id="4x8_grid")], indirect=True)
-def test_visualize_tensor_2d_sharded(mesh_device):
-    rows, cols = mesh_device.shape
-    tile_size = 32
-    full_tensor = torch.rand((1, 1, tile_size * rows, tile_size * cols), dtype=torch.bfloat16)
-    mesh_mapper = ttnn.create_mesh_mapper(
-        mesh_device,
-        ttnn.MeshMapperConfig(
-            placements=[
-                ttnn.PlacementShard(2),
-                ttnn.PlacementShard(3),
-            ],
-        ),
-    )
-    ttnn_tensor = ttnn.from_torch(full_tensor, mesh_mapper=mesh_mapper, layout=ttnn.Layout.ROW_MAJOR)
-    ttnn.visualize_tensor(ttnn_tensor)
-    ttnn_tensor = ttnn_tensor.to(mesh_device)
-    ttnn.visualize_tensor(ttnn_tensor)
-
-
-@pytest.mark.parametrize("mesh_device", [pytest.param((4, 8), id="4x8_grid")], indirect=True)
-def test_visualize_tensor_row_major_distribution(mesh_device):
-    rows, cols = mesh_device.shape
-    tile_size = 32
-    full_tensor = torch.rand((1, 1, tile_size * rows, tile_size * cols), dtype=torch.bfloat16)
-    mesh_mapper = ttnn.create_mesh_mapper(
-        mesh_device,
-        ttnn.MeshMapperConfig(
-            placements=[ttnn.PlacementReplicate(), ttnn.PlacementShard(3)],
-            mesh_shape_override=ttnn.MeshShape(2, 16),
-        ),
-    )
-    ttnn_tensor = ttnn.from_torch(full_tensor, mesh_mapper=mesh_mapper, layout=ttnn.Layout.ROW_MAJOR)
-    ttnn.visualize_tensor(ttnn_tensor)
-    ttnn_tensor = ttnn_tensor.to(mesh_device)
-    ttnn.visualize_tensor(ttnn_tensor)
-
-
-@pytest.mark.parametrize("mesh_device", [pytest.param((4, 8), id="4x8_grid")], indirect=True)
-def test_visualize_tensor_submesh_distribution(mesh_device):
-    rows, cols = mesh_device.shape
-    tile_size = 32
-    full_tensor = torch.rand((1, 1, tile_size * rows, tile_size * cols), dtype=torch.bfloat16)
-    mesh_mapper = ttnn.create_mesh_mapper(
-        mesh_device,
-        ttnn.MeshMapperConfig(
-            placements=[ttnn.PlacementReplicate(), ttnn.PlacementShard(3)],
-            mesh_shape_override=ttnn.MeshShape(3, 4),
-        ),
-    )
-    ttnn_tensor = ttnn.from_torch(full_tensor, mesh_mapper=mesh_mapper, layout=ttnn.Layout.ROW_MAJOR)
-    ttnn.visualize_tensor(ttnn_tensor)
-    ttnn_tensor = ttnn_tensor.to(mesh_device)
-    ttnn.visualize_tensor(ttnn_tensor)
-
-
 def rms_norm(x, gamma, eps):
     return x * torch.rsqrt(x.pow(2).mean(-1, keepdim=True) + eps) * gamma
